MFC: ntp 4.1.0.
	See ChangeLog for all the many changes/bug fixes/new drivers.

	This code has been in CURRENT for months and several people use
	this version on 4.4-STABLE.

Approved by:	re

+++ contrib/ntp/util/hist.c
@@ -8,13 +8,13 @@
  */
 
 #ifdef HAVE_CONFIG_H
-#include <config.h>
+# include <config.h>
 #endif
-#include <sys/time.h>
-#include <stdio.h>
-#include <stdlib.h>
 
 #include "ntp_types.h"
+
+#include <stdio.h>
+#include <stdlib.h>
 
 #define NBUF 100001		/* size of basic histogram */
 #define NSRT 20000		/* size of overflow histogram */
@@ -79,7 +79,7 @@ main(
 	}
 	if (n == 0)
 	    return;
-	qsort((char *)ovfl, (int)n, sizeof(long), col);
+	qsort((char *)ovfl, (size_t)n, sizeof(long), col);
 	w = 0;
 	j = 0;
 	for (i = 0; i < n; i++) {
